[WIP] send ensure result before bulk of migration protocol
